+++ app/adapters/sqlite.py
@@ -146,7 +146,7 @@ class SqliteDatabase(object):
 	"""This adapter gives access to each of the SQLite database classes that
 	are to be used by the app for data mutation an querying.
 
-	Methods: get_uowm, get_view
+	Methods: set_up, get_uowm, get_view
 	"""
 	def __init__(self, cfg: dict):
 		"""SqliteDatabase's constructor.